avbtool: Add --format to calculate_vbmeta_digest
Add an optional flag controlling the output format of the command. For
now, only support 'hex', which defaults to the previous behavior.

In prevision of a future patch modifying the output format, make the
output file a binary file, instead of a text file. This shouldn't affect
the current behavior as the written hex string is simple enough to be
generated in binary mode with no change.

Test: atest libavb_host_unittest
Flag: EXEMPT refactor
Change-Id: Ie5ee1e283ad85fb0aefd386b19dfb9a53b1cf075