avbtool: Fix crashes when writing to stdout
Using FileType('wb...') for output files passed as CLI arguments opens
them as binary file objects but, because of default=sys.stdout, they
default to text file objects (type of sys.stdout) when the CLI argument
is missing.

As a result and because the write() method of binary and text file
objects respectively take bytes and str parameters, avbtool can only
pass one of the two types and therefore panics when the other is used.

For example:

  $ avbtool generate_test_image --image_size 10
  Traceback (most recent call last):
    File "<frozen runpy>", line 198, in _run_module_as_main
    [...]
    File "avbtool.py", line 2225, in generate_test_image
  TypeError: write() argument must be str, not bytearray

Instead, as the argparse.FileType documentation guarantees that

> FileType objects understand the pseudo-argument '-' and automatically
> convert this into [...] sys.stdout for writable FileType objects

change all calls to default to '-', making FileType('b...') create the
expected binary files when defaulting to the standard output.

For consistency, also change users of FileType('t...').

Fixes: 436281780
Flag: EXEMPT bugfix
Test: avbtool generate_test_image --image_size 10
Change-Id: I58f40dd92bc9fb31fbb3e56ac77beab918259d5d