Refactor all key operations into RSAPublicKey
Move the RSA verification logic into the RSAPublicKey class to keep the
algorithm details hidden from the caller.

This results in multiple ways to initialize the class so named
constructor methods are introduced that delegate to a common, low level
__init__().

When initializing from an AVB public key blob, a temporary file
containing a DER copy of the key is created. The class is updated to act
as a context manager so that the temporary file is deleted when it's not
longer needed. Callers are updated to use the context manager pattern.

Test: atest libavb_host_unittest
Bug: 412315350
Change-Id: Ic4651f9863d9fc32d2bba3a5c946ee2ec8dda87d