Rust avb: add mismatched_lifetime_syntaxes lint
This lint isn't part of the Android standard set, but only because there
are too many violations to readily fix. It's a useful lint though, in
particular it makes it clearer when return values carry lifetimes.
Previously some of these lifetimes were elided, which can cause
surprising results since a user might naturally assume there is no
borrow involved.

This CL fixes the violations and enables the lint in the top-level file
for each build target to prevent against future violations.

Bug: b/456142394
Test: presubmit
Change-Id: I7f12671f867c7bd2032e348b9d9f3763881a13eb